+++ retz-server/src/main/java/io/github/retz/scheduler/Launcher.java
@@ -20,6 +20,7 @@ import com.j256.simplejmx.server.JmxServer;
 import io.github.retz.db.Database;
 import io.github.retz.mesosc.MesosHTTPFetcher;
 import io.github.retz.protocol.data.Job;
+import io.github.retz.web.StatusCache;
 import io.github.retz.web.WebConsole;
 import org.apache.commons.cli.*;
 import org.apache.mesos.Protos;